backend/requirements.txt
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.13.3
black==25.1.0
bs4==0.0.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.0
defusedxml==0.7.1
Django==5.1.1
django-filter==25.1
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
drf-extra-fields==3.7.0
drf-spectacular==0.29.0
et_xmlfile==2.0.0
fake-useragent==2.1.0
filetype==1.2.0
flake8==7.1.1
gunicorn>=21.0.0
h11==0.14.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.5
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lxml==5.3.2
MarkupSafe==3.0.2
mccabe==0.7.0
mypy_extensions==1.1.0
numpy==2.2.4
oauthlib==3.2.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.4.0
pluggy==1.5.0
psycopg2-binary==2.9.11
py==1.11.0
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.10.1
pyparsing==3.2.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
PySocks==1.7.1
pytest==8.3.3
pytest-django==4.9.0
pytest-pythonpath==0.7.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.28.0
ruff==0.13.0
selenium==4.30.0
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.2
toml==0.10.2
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
useragent==0.1.1
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
